backend/requirements.txt
# Core FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis for caching and pub/sub
redis==5.0.1
aioredis==2.0.1

# HTTP client for external APIs
httpx==0.25.2

# System monitoring
psutil==5.9.8

# Data validation and processing
email-validator==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI

# Code quality
black==23.11.0
isort==5.12.0
ruff==0.1.6
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Optional: For future features
# celery==5.3.4  # Background tasks
# flower==2.0.1  # Celery monitoring
# sentry-sdk==1.38.0  # Error tracking
# python-socketio==5.10.0  # WebSocket support (if needed beyond FastAPI)
# Export functionality
openpyxl==3.1.2
reportlab==4.0.8
